Event timing normalization caused crashes

+++ GBAEmulator/Scheduler/Scheduler.Event.cs
@@ -1,16 +1,14 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace GBAEmulator.Scheduler
 {
     public class Event : IComparable<Event>
     {
-        public int Time;
+        public long Time;
         public delegate void Handler(Event sender, Scheduler scheduler);
         private readonly Handler _Handler;
 
-        public Event(int Time, Handler _Handler)
+        public Event(long Time, Handler _Handler)
         {
             this.Time = Time;
             this._Handler = _Handler;
@@ -23,7 +21,7 @@ namespace GBAEmulator.Scheduler
 
         public int CompareTo(Event other)
         {
-            return this.Time - other.Time;
+            return (int)(this.Time - other.Time);
         }
 
     }
